djangoapi/postapi/requirements.txt
asgiref==3.11.1
Django==6.0.3
djangorestframework==3.17.0
psycopg2-binary==2.9.11
sqlparse==0.5.5
tzdata==2025.3
